+++ src/rest-api/src/features/members/extractors/member_service.rs
@@ -24,7 +24,7 @@ impl FromRequestParts<AppState> for service::members::MemberService {
         };
 
         Ok(Self::new(
-            Arc::new(state.db.clone()),
+            state.db.clone(),
             Arc::new(state.server_ctl.clone()),
             Arc::new(xmpp_service),
             ConcurrentTaskRunner::default(),