+++ src/rest-api/src/main.rs
@@ -43,7 +43,9 @@ async fn main() {
         let figment = AppConfig::figment();
         let ref log_config = match figment.extract_inner::<LogConfig>("log") {
             Ok(config) => config,
-            Err(err) if err.missing() => Default::default(),
+            Err(err) if err.missing() => {
+                unreachable!("Defaults should be provided by `AppConfig::figment`")
+            }
             Err(err) => panic!("Invalid `log` config: {err:?}"),
         };
         init_subscribers(log_config)
@@ -129,7 +131,7 @@ async fn run(
         Err(_) if lifecycle_manager.is_restarting() => {
             warn!("The Prose Pod API is missing some static configuration. Serving only utility routes.");
 
-            let addr = SocketAddr::new(defaults::api_address(), defaults::api_port());
+            let addr = SocketAddr::new(defaults::API_ADDRESS, defaults::API_PORT);
 
             let app = factory_reset_router(&minimal_app_state);
 
@@ -180,9 +182,12 @@ async fn startup(app_config: AppConfig, minimal_app_state: MinimalAppState) -> R
 
 #[instrument(level = "trace", skip_all)]
 async fn init_dependencies(app_config: AppConfig, base: MinimalAppState) -> AppState {
-    let db = db_conn(&app_config.api.databases.main)
-        .await
-        .expect("Could not connect to the database.");
+    let db = db_conn(
+        &app_config.api.databases.main_read,
+        &app_config.api.databases.main_write,
+    )
+    .await
+    .expect("Could not connect to the database.");
 
     let license_service_impl = match LiveLicenseService::from_config(&app_config) {
         Ok(service) => service,