+++ src/rest-api/tests/prelude/mocks/mock_server_ctl.rs
@@ -10,9 +10,9 @@ use service::{
     auth::AuthToken,
     errors::Forbidden,
     members::MemberRole,
+    models::DatabaseRwConnectionPools,
     prosody::{prosody_bootstrap_config, AsProsody as _, ProsodyConfig},
     prosody_config_from_db,
-    sea_orm::DatabaseConnection,
     util::either::{Either, Either3},
     xmpp::{
         server_ctl::{self, errors::GroupNotFound, Error},
@@ -23,10 +23,10 @@ use service::{
 
 use std::sync::{Arc, RwLock};
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 pub struct MockServerCtl {
     pub(crate) state: Arc<RwLock<MockServerCtlState>>,
-    pub(crate) db: DatabaseConnection,
+    pub(crate) db: DatabaseRwConnectionPools,
 }
 
 #[derive(Debug)]
@@ -44,7 +44,7 @@ pub struct MockServerCtlState {
 }
 
 impl MockServerCtl {
-    pub fn new(state: Arc<RwLock<MockServerCtlState>>, db: DatabaseConnection) -> Self {
+    pub fn new(state: Arc<RwLock<MockServerCtlState>>, db: DatabaseRwConnectionPools) -> Self {
         Self { state, db }
     }
 
@@ -82,7 +82,8 @@ impl ServerCtlImpl for MockServerCtl {
         self.check_online()?;
 
         let new_config =
-            prosody_config_from_db(&self.db, app_config, Some(server_config.to_owned())).await?;
+            prosody_config_from_db(&self.db.read, app_config, Some(server_config.to_owned()))
+                .await?;
 
         let mut state = self.state.write().unwrap();
         state.applied_config = Some(Arc::new(new_config));