+++ src/service/src/features/auth/util.rs
@@ -3,7 +3,6 @@
 // Copyright: 2024–2025, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
-use rand::{distributions::Alphanumeric, thread_rng, Rng as _};
 use secrecy::SecretString;
 
 /// Generates a random secret string.
@@ -11,13 +10,7 @@ use secrecy::SecretString;
 pub fn random_secret(length: usize) -> SecretString {
     assert!(length >= 16);
 
-    // NOTE: Code taken from <https://rust-lang-nursery.github.io/rust-cookbook/algorithms/randomness.html#create-random-passwords-from-a-set-of-alphanumeric-characters>.
-    thread_rng()
-        .sample_iter(&Alphanumeric)
-        .take(length)
-        .map(char::from)
-        .collect::<String>()
-        .into()
+    crate::util::random_string_alphanumeric(length).into()
 }
 
 /// Generates a random secret string (URL-safe).
@@ -25,4 +18,23 @@ pub fn random_secret(length: usize) -> SecretString {
 pub fn random_secret_url_safe(length: usize) -> SecretString {
     // NOTE: Already URL-safe.
     random_secret(length)
+}
+
+#[must_use]
+#[inline]
+pub fn random_oauth2_registration_key() -> SecretString {
+    use rand::RngCore as _;
+
+    let mut key = [0u8; 256];
+    rand::thread_rng().fill_bytes(&mut key);
+
+    // fn bytes_to_hex(bytes: &[u8]) -> String {
+    //     bytes.iter().map(|byte| format!("{:02x}", byte)).collect()
+    // }
+    fn bytes_to_base64(bytes: &[u8]) -> String {
+        use base64::Engine as _;
+        base64::prelude::BASE64_STANDARD.encode(bytes)
+    }
+
+    SecretString::from(bytes_to_base64(&key))
 }