+++ CHANGELOG.md
@@ -9,7 +9,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- WARN: Do not move the next line and add changelog entries **under** it.
        It’s used by `task release` when updating the changelog. -->
-[Unreleased]: https://github.com/prose-im/prose-pod-api/compare/v0.19.1...HEAD
+[Unreleased]: https://github.com/prose-im/prose-pod-api/compare/v0.19.2...HEAD
+
+## [0.19.2] - 2025-09-23
+
+[0.19.2]: https://github.com/prose-im/prose-pod-api/compare/v0.19.1...v0.19.2
+
+### Changed
+
+- chore(app-config): Use `figment` correctly instead of using `serde(default)` (in `4b1c13b7`)
+
+### Added
+
+- test(app-config): Add tests for smart default values (in `fd96cf88`)
+
+### Fixed
+
+- fix: Fix database locks by opening two connection pools (in `8d808920`)
+- fix(local-run): Fix `task telemetry:start` not working when Docker network changed (in `cf3c6b46`)
 
 ## [0.19.1] - 2025-09-16